HSEARCH-3115 Remove the unnecessary abstraction over analysis definition registries
We only have one implementation of each interface, and in the
Elasticsearch case we even rely on methods declared in the
implementation, so there's no point exposing interfaces.
